Added ability to ignore directories by name
Although this works with no obvious bugs, it is a mess. All of the
command line argument parsing logic is stuffed into rin.cpp, and
a vector for the list of directory names to be ignore has been added to
several functions. Eventually, the command line argument parsing will be
in a different header and cpp file, and a class will be created to
pass one object to functions instead of several arguments.